requirements.txt
astroid==2.2.5
atomicwrites==1.2.1
attrs==18.2.0
isort==4.3.17
Jinja2==2.10.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==4.3.0
mysql-connector-python==8.0.13
pluggy==0.8.0
protobuf==3.6.1
py==1.7.0
pylint==2.3.1
pytest==4.0.2
python-dateutil==2.7.5
python-slugify==2.0.0
pytz==2019.1
six==1.12.0
typed-ast==1.3.4
Unidecode==1.0.23
wrapt==1.11.1
